+++ packages/client/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @eudi-verify/client
 
+## 1.3.2
+
 ## 1.3.1
 
 ### Patch Changes